requirements.txt
pandas
numpy
scipy
matplotlib
plotly
python-binance
TA-Lib
kaleido